feat: Settings file refactor
