+++ airsync-mac/Screens/Settings/Components/SettingsHeaderView.swift
@@ -1,0 +1,18 @@
+import SwiftUI
+
+struct SettingsHeaderView: View {
+    let title: String
+    let icon: String
+    
+    var body: some View {
+        HStack(spacing: 8) {
+            Image(systemName: icon)
+                .foregroundStyle(Color.accentColor)
+            Text(title)
+                .font(.system(size: 13, weight: .bold))
+                .foregroundStyle(.secondary)
+            Spacer()
+        }
+        .padding(.horizontal, 8)
+    }
+}
